+++ src/main/java/org/vietsearch/essme/controller/RequestResponseController.java
@@ -1,5 +1,6 @@
 package org.vietsearch.essme.controller;
 
+import com.google.firebase.auth.FirebaseAuthException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -8,6 +9,7 @@ import org.springframework.data.mongodb.core.query.TextCriteria;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
+import org.vietsearch.essme.filter.AuthenticatedRequest;
 import org.vietsearch.essme.repository.RequestResponseRepository;
 import org.vietsearch.essme.model.request_response.*;
 import org.vietsearch.essme.repository.direct_request.DirectRequestRepository;
@@ -16,6 +18,7 @@ import javax.validation.Valid;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/api/requests")
@@ -26,6 +29,7 @@ public class RequestResponseController {
     @Autowired
     private DirectRequestRepository directRequestRepository;
 
+
     @GetMapping
     public Page<Request> getRequests(@RequestParam(value = "page", defaultValue = "0") int page, @RequestParam(value = "size", defaultValue = "20") int size, @RequestParam(value = "sort", defaultValue = "createdAt") String sortAttr, @RequestParam(value = "desc", defaultValue = "false") boolean desc) {
         Sort sort = Sort.by(sortAttr);
@@ -35,6 +39,7 @@ public class RequestResponseController {
         Page<Request> requestPage = requestRepository.findAll(PageRequest.of(page, size, sort));
         return requestPage;
     }
+
 
     @GetMapping("/{id}")
     public Request getRequestById(@PathVariable("id") String id) {
@@ -59,8 +64,16 @@ public class RequestResponseController {
 
     @PutMapping("/{id}")
     @ResponseStatus(HttpStatus.OK)
-    public Request updateRequest(@PathVariable("id") String id, @Valid @RequestBody Request request) {
+
+    public Request updateRequest(AuthenticatedRequest authenticatedRequest,@PathVariable("id") String id, @Valid @RequestBody Request request) {
+        String uuid = authenticatedRequest.getUserId();
         if (requestRepository.existsById(id)) {
+            // check
+            if(!matchUserRequest(uuid, id)){
+                throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Permission denied", null);
+            }
+
+            // update
             request.set_id(id);
             requestRepository.save(request);
             return request;
@@ -71,8 +84,15 @@ public class RequestResponseController {
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.OK)
-    public String deleteRequest(@PathVariable("id") String id) {
+    public String deleteRequest(AuthenticatedRequest authenticatedRequest,@PathVariable("id") String id) {
+        String uuid = authenticatedRequest.getUserId();
         if (requestRepository.existsById(id)) {
+            // check
+            if(!matchUserRequest(uuid, id)){
+                throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Permission denied", null);
+            }
+
+            // delete
             requestRepository.deleteById(id);
             return "Deleted";
         } else {
@@ -82,10 +102,11 @@ public class RequestResponseController {
 
     @PostMapping("/{requestId}/responses")
     @ResponseStatus(HttpStatus.CREATED)
-    public Request addResponse(@PathVariable("requestId") String requestId, @Valid @RequestBody Response response) {
+    public Request addResponse(AuthenticatedRequest authenticatedRequest, @PathVariable("requestId") String requestId, @Valid @RequestBody Response response) {
         Request request = requestRepository.findById(requestId).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Request not found", null));
         if (request.getResponses() == null)
             request.setResponses(new ArrayList<>());
+        response.setUid(authenticatedRequest.getUserId());
         request.getResponses().add(response);
         return requestRepository.save(request);
     }
@@ -98,17 +119,20 @@ public class RequestResponseController {
 
     @PutMapping("/{requestId}/responses/{responsesId}")
     @ResponseStatus(HttpStatus.OK)
-    public Response updateResponse(@PathVariable("requestId") String requestId, @PathVariable("responsesId") String responsesId, @Valid @RequestBody Response response) {
+
+    public Response updateResponse(AuthenticatedRequest authenticatedRequest,@PathVariable("requestId") String requestId, @PathVariable("responsesId") String responsesId, @Valid @RequestBody Response response) {
+        String uuid = authenticatedRequest.getUserId();
         Request request = requestRepository.findById(requestId).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Request not found", null));
         if (request.getResponses() != null) {
             for (Response res : request.getResponses()) {
-                if (res.get_id().equals(responsesId)) {
+                if (matchUserResponse(uuid, responsesId, res)) {
                     res.setExpertId(response.getExpertId());
                     res.setContent(response.getContent());
                     res.setUpdatedAt(new Date());
                     requestRepository.save(request);
                     return res;
                 }
+                else throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Permission denied", null);
             }
         }
         throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Response not found", null);
@@ -116,19 +140,36 @@ public class RequestResponseController {
 
     @DeleteMapping("/{requestId}/responses/{responsesId}")
     @ResponseStatus(HttpStatus.OK)
-    public String deleteAnswer(@PathVariable("requestId") String requestId, @PathVariable("responsesId") String responseId) {
+    public String deleteResponse(AuthenticatedRequest authenticatedRequest, @PathVariable("requestId") String requestId, @PathVariable("responsesId") String responseId) {
+        String uuid = authenticatedRequest.getUserId();
         Request request = requestRepository.findById(requestId).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Request not found", null));
         if (request.getResponses() != null) {
             for (Response res : request.getResponses()) {
-                if (res.get_id().equals(responseId)) {
+                if (matchUserResponse(uuid, responseId, res)) {
                     request.getResponses().remove(res);
                     requestRepository.save(request);
                     return "Deleted";
                 }
+                else throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Permission denied", null);
             }
         }
         throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Response not found", null);
     }
+
+
+    private boolean matchUserRequest(String uuid, String requestId){
+        // return true if uuid created request
+        Optional<Request> optional = requestRepository.findById(requestId);
+        return optional.map(res -> res.getCustomerId().equals(uuid)).orElse(false);
+    }
+
+    private boolean matchUserResponse(String uuid, String responseChangedId, Response response){
+        // return true if uuid created response
+        if(!response.get_id().equals(responseChangedId))
+            return false;
+        return response.getExpertId().equals(uuid);
+    }
+
 
     @GetMapping("/direct/{requestId}")
     @ResponseStatus(HttpStatus.OK)
@@ -175,4 +216,5 @@ public class RequestResponseController {
         directRequestRepository.deleteById(id);
         return "Deleted request: " + id;
     }
+
 }
